Update index_2027.md
Cody,
Your resume looks really good.  I would just verify the date on line 53 it says 2027.  Also I would fix a few of the words that are capitalized, for example Application and Detailed.  I would also suggest to spell out BS.  I also wondered if Related Coursework was really coursework or if it was jobs.